Added new particle/Bug fixes
Added a new particle that plays when the player dashes there is a small burst of while pixels then the players is moved forward. Found and fixed a bug that allows the player to jump on hazard objects that should destroy them, the fix was simply making the players jump value 0 as soon as they collide with the hazard.