Level design/New art/New feature
made and added a new conveyor belt asset that I attached a platform effector to, when the player lands on one of the conveyor belt tiles a force is applied to a certain direction. Added a new checkpoint system that allows the player to keep some progress while traversing levels. Added checkpoints to level 1, 2,and 3. Added a new section to level 3 that is challenging and makes use of the new conveyor belt I imported.