Fixed checkpoints/Small level changes
Fixed a problem where my checkpoints would not work on any other level than the first level, to fix it I Simply had to add the player pos tracker to the player as I forgot to add it. Fixed a problem with the conveyor belt where the players velocity was being set to 0 by a ground check. Added another section to level 3.